Fix bug in Mac toolbar customization (#36)
There was a bug in the saving of Mac toolbar content customization where the content of the data viewer toolbar would be lost when customizing the file selector toolbar in a session where the data viewer wasn't opened. Fixed the settings persistence to only save changes to the toolbar which is currently displayed.

Also updated the Mac SDK to the version required for Qt 5.15, which I plan to use for the next release. I just tested that PortaBase successfully builds and runs with it, although I'll want to fix a cosmetic issue or two before releasing.